+++ backend/src/main/java/com/portfolio/weatheralert/WeatherAlertApplication.java
@@ -6,13 +6,11 @@ import org.springframework.boot.context.properties.ConfigurationPropertiesScan;
 import org.springframework.cache.annotation.EnableCaching;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 import org.springframework.scheduling.annotation.EnableAsync;
-import org.springframework.scheduling.annotation.EnableScheduling;
 
 @ConfigurationPropertiesScan
 @EnableAsync
 @EnableCaching
 @EnableJpaAuditing
-@EnableScheduling
 @SpringBootApplication
 public class WeatherAlertApplication {